(style) Experimenting with an extremely unorthodox assignment style…
I think I’ll stick with this, I tend to like it. It has several benefits: I think it keeps the code clearer, once you learn to focus on the code to the left; it *also* makes it quite easy to scan the code for things that are ‘provided’ for use in the outside world. Every single function or value that is exposed for instances of one of these ‘classes’ has an immediate presence on the right-hand-margin.