(fix) Completely re-implemented `Acquisition.dispatch()`.
There were quite a few inconsistencies and bugs in the whole code-path surrounding `dispatch()` and
`complete()`, not to mention that they were pointlessly synchronous.

All `dispatch()`es for a given query are now, well, ‘dispatched’ asynchronously.

(Also, several small bugs in the `exports` and `imports` system were squashed in the process.)