(meta new api incomplete) Preforming meta-project set-up.
I've added a `package.json` for NPM, and set up some tasks for Grunt.js.

I'm preparing the project for a time when I'll be providing a prepared “distribution” file to be
included into the web-client (I have every intention of providing a [Bower][] component for
`from()`).

Finally, I started adding Grunt integration specifically for running the various test combinations;
but, unfortunately, there's no `grunt-testling`; and calling out to the shell by hand is
excruciatingly difficult in Node / Grunt. I can't be arsed, so I'll wait either for a
`grunt-testling`, or at least for `grunt-exec` to be updated for Grunt 0.4.0a.

  [Bower]: <http://twitter.github.com/bower/> "The Browser Package Manager"