(- new plumbing API) `subclass()` can now pass `arguments` intact
Although passing `this` and each of the arguments on to an internal function (via a pattern like
`fn.apply(this, arguments)`) is usually sufficient, ocassionally one also wants access to something
more (for instance, `arguments.callee` or `arguments.callee.caller`). By passing `true` as the last
argument to `subclass()`, you can now have the `arguments` object passed wholesale to the inner
function (the one used as the `runtimeBody`.

This is most effective if you name the only argument to the inner function, itself, as `arguments`
(although I'm not sure how widely supported an argument named `arguments` might be, since that's
theoretically overriding the `arguments` object itself):

    var inner = function(arguments){
       // ... do something with `arguments.caller`
    }
    subclass(Function, Something, inner, true)